refactor(commit_round): split the commit-validation service out of apply
Move staging, sender authorization, and action-matching (the RFC "commit validation service") into commit_round::validate, leaving apply.rs as the best-first loop and its local/remote apply paths. Share leave_or_update between the paths, and use plain pub (gated by the private modules) instead of pub(super) throughout.
